Crate minhook

The minhook-rs library

This library provides function hooking support to Rust by providing a wrapper around the MinHook library.

Modules

function
Module containing information about hookable functions.

panic
Panic handling for panics caught at foreign code boundaries in detour functions.

Macros

static_hooks!
Defines one or more static hooks.

Structs

Hook
A hook that is destroyed when it goes out of scope.

HookQueue
A queue of hook changes to be applied at once.

StaticHook
A hook with a static lifetime.

StaticHookWithDefault
A hook with a static lifetime and a default detour closure.

Enums

Error
The error type for all hooking operations.

FunctionName
A function name used for dynamically hooking a function.

Functions

uninitialize
Uninitializes the minhook-rs library.

Type Definitions

Result
Result type for most functions and methods in this module.