Struct minhook::Hook

pub struct Hook<T: Function> {
    // some fields omitted
}

A hook that is destroyed when it goes out of scope.

Methods

impl<T: Function> Hook<T>

unsafe fn new<D>(target: T, detour: D) -> Result<Hook<T>> where T: HookableWith<D>, D: Function

Create a new hook given a target function and a compatible detour function.

The hook is disabled by default. Even when this function is succesful, there is no guaranteee that the detour function will actually get called when the target function gets called. An invocation of the target function might for example get inlined in which case it is impossible to hook at runtime.

Safety

This function is unsafe because there are a few guarantees to be upheld by the caller that can not be checked:

The target and detour function pointers must point to valid memory for the entire lifetime of this hook.

The given target function type must uniquely match the actual target function. This means two things: the given target function type has to be correct, but also there can not be two function pointers with different signatures pointing to the same code location. This last situation can for example happen when the Rust compiler or LLVM decide to merge multiple functions with the same code into one.

fn trampoline(&self) -> T::Unsafe

Returns an unsafe reference to the trampoline function.

Calling the returned function is always unsafe as it will point to invalid memory after the hook is destroyed.

fn enable(&self) -> Result<()>

Enables this hook.

Consider using a HookQueue if you want to enable/disable a large amount of hooks at once.

fn disable(&self) -> Result<()>

Disables this hook.

Consider using a HookQueue if you want to enable/disable a large amount of hooks at once.