Struct minhook::StaticHookWithDefault

pub struct StaticHookWithDefault<T: Function> {
    // some fields omitted
}

A hook with a static lifetime and a default detour closure.

This hook can only be constructed using the static_hooks! macro. It has one of the following forms:

#[ATTR]* pub? impl HOOK_VAR_NAME for PATH::TO::TARGET: FN_TYPE = CLOSURE_EXPR;
#[ATTR]* pub? impl HOOK_VAR_NAME for "FUNCTION" in "MODULE": FN_TYPE = CLOSURE_EXPR;

Before accessing this hook it is required to call initialize() once. Accessing the hook before initializing or trying to initialize the hook twice (even after the first attempt failed) will result in a panic.

Methods

impl<T: Function> StaticHookWithDefault<T>

unsafe fn initialize(&self) -> Result<()>

Initialize and install the underlying hook.

Panics

Panics if the hook was accessed or initialized before this call.

Safety

See documentation for Hook::create() and Hook::create_api()